Time: 0 ms (100%), Space: 11.7 MB (95.54%) - LeetHub

+++ 1319-unique-number-of-occurrences/1319-unique-number-of-occurrences.cpp
@@ -2,13 +2,13 @@ class Solution {
 public:
     bool uniqueOccurrences(vector<int>& arr) {
         sort(arr.begin(), arr.end());
-        int arrSize = arr.size();
+        unsigned short arrSize = arr.size();
 
-        int prevNum = arr[0];
-        int numCount = 1;
+        short prevNum = arr[0];
+        unsigned short numCount = 1;
         vector<bool> checkNumCount(1001);
 
-        for (int i = 1; i < arrSize; ++i)
+        for (unsigned short i = 1; i < arrSize; ++i)
         {
             if (arr[i] == prevNum)
             {
